[Quirks] Move the Quirk table to it's own file https://bugs.webkit.org/show_bug.cgi?id=322674 rdar://185947168
Reviewed by Brent Fulgham.

This patch moves the Quirk table to it's own file.

Test: Tools/TestWebKitAPI/Tests/WebCore/Quirks.cpp

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/QuirkMatch.cpp:
(WebCore::Quirk::apply const): Deleted.
* Source/WebCore/page/QuirkMatch.h:
(WebCore::QuirkBehaviors::QuirkBehaviors): Deleted.
* Source/WebCore/page/QuirkTable.cpp: Added.
(WebCore::Quirk::apply const):
(WebCore::resolveSiteSpecificQuirks):
* Source/WebCore/page/QuirkTable.h: Added.
(WebCore::QuirkBehaviors::QuirkBehaviors):
* Source/WebCore/page/Quirks.cpp:
* Tools/TestWebKitAPI/Tests/WebCore/Quirks.cpp:

Canonical link: https://commits.webkit.org/319930@main
