+++ Source/WebCore/page/QuirkTable.h
@@ -1,0 +1,61 @@
+/*
+ * Copyright (C) 2026 Apple Inc. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS''
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS
+ * BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
+ * THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#include <WebCore/QuirkMatch.h>
+#include <WebCore/QuirksData.h>
+#include <initializer_list>
+#include <optional>
+
+namespace WebCore {
+
+class QuirkBehaviors {
+public:
+    constexpr QuirkBehaviors() = default;
+
+    consteval QuirkBehaviors(std::initializer_list<QuirksData::SiteSpecificQuirk> quirks)
+    {
+        for (auto quirk : quirks)
+            m_bits.set(static_cast<size_t>(quirk));
+    }
+
+    constexpr const QuirksData::QuirkBitSet& bits() const LIFETIME_BOUND { return m_bits; }
+
+private:
+    QuirksData::QuirkBitSet m_bits;
+};
+
+struct Quirk {
+    QuirkMatch match;
+    QuirkBehaviors behaviors { };
+    std::optional<QuirkSite> site { };
+
+    void apply(QuirksData&) const;
+};
+
+WEBCORE_EXPORT QuirksData resolveSiteSpecificQuirks(const QuirkMatchContext&);
+
+} // namespace WebCore
